Refactor layout and cleanup NavMenu CSS
This commit removes the redundant and redundant NavMenu.razor.css, restructures the NavMenu component, and changes the location of NavMenu in layout. The changes have been made to become more aligned with bootstrap standards and improve the look and feel. Also, the CSS and scripts path are updated in App.razor.