Refactor service registration and update project hierarchy
This commit introduces refactored service registration to improve the organization of the project. All service configurations now reside in dedicated methods in the `ServiceCollectionExtensions` class under `RegisterServices` namespace for cleaner codebase and ease of maintenance. It also includes the invoked methods for service registration in Program.cs and the necessary files/tests for the service configurations. Project hierarchy has been updated as well, with the addition of new files and the renaming of existing ones for better organization. Changes have also been made to `Home.razor` and various `.xml` files.