Update test paths in workflow file
The GitHub workflow file dotnet.yml was updated to change the working directory paths for unit, integration and BUnit tests. The correction aligns with the renaming of the source directory from 'BlazorBlogs' to 'BlazorBlog'. Moreover, the project solution file 'mpaulosky_BlogAppV3.sln' was updated to include references to various workflow files in the '.github' directory.