Remove 'todos-to-issues' workflow and update .NET setup
Removed the 'todos-to-issues' workflow from the GitHub Actions setup and deleted references to it in the solution file. Updated the .NET setup in the 'dotnet.yml' workflow to use a 'global.json' file instead of specifying version directly. Also adjusted the file paths in the 'run' commands in the unit tests, integration tests, and BUnit tests sections. Lastly, cleaned up code in 'Program.cs' by removing unused using statements and a TODO comment.