Update Qodana Github workflow configuration
The configuration for the Qodana code quality checker in the GitHub workflow has been updated. Changes include a switch to an older version of actions/checkout, the addition of an explicit Qodana action, and outputting reports in HTML format. Furthermore, delete the permissions section and upload the Qodana report as an artifact at the end of the process.